resume-ai-backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-community
langchain-google-genai
python-docx
PyMuPDF
jinja2
pydantic
httpx
starlette
